Update rdc_importer.py
Adjusted the logic to instead pull back texture names as diffuse or normal etc, this should give insight into whether the textures are correct and incorrect before starting the match process. If the textures look incorrect, we are likely pulling from the wrong draw actions and need to review and change those values.