Update __init__.py for new feature
Added in a new text property to support the inclusion of multiple ranges in the render doc import. Doing this will allow us to provide multiple ranges so we can skip unnecessary actions, and import only what we need to. 

Due to how AC draws now, we can pick and choose which draw layers we are importing from without needing to bring in multiple duplicate meshes.